e2e: serial: schedstall: add tests with jobs
The default scheduler framework setting watches pod not in terminal
phase. On the node, the podresources endpoint does report also pods in
terminal phase, up until are deleted. This mismatch creates differences
in PFP value, so the scheduler never reconciles, leading to a so-called
"scheduler stall".

While workaround were provided (podExcludes), noone was deemed
conclusive. The implemented for the short/mid term is to just have
another separate independent informer watching pods in all states, just
and only to serve the NRT scheduler plugins.

In this change we add tests to trigger scheduler stalls and exercise the
code, making sure stalls are avoided.
The easiest way to create undeleted pods in terminal phase is to use
jobs, so this is what we do.

Note this group of tests intentionally duplicates existing tests, adding
jobs to the node conditions before to run them.
The alternative is to add yet another column in the test matrix for
the existing case, which is not obviously better considering the purpose
of these tests.

Signed-off-by: Francesco Romani <fromani@redhat.com>