e2e: verify nrt data after each test
Following #598, #608 stabilize the e2e tests even more by verifying after each test that the NRT data is settled to the initial. This will help in troubleshooting failures related to to unclean NRT data and in a meaninful cooldown.

PS: we assume that the cluster is clean before running the suite, as in it doesn't have workloads apart from the system and setup workloads that are needed for executing serial. Also, we assume no sudden interrupts, resources consumptions/releases during the run, which was also the case before this update.

Signed-off-by: shereenH <shajmakh@redhat.com>