e2e: introduce E2E_NROP_COOLDOWN_THRESHOLD
In order to verify that NRT data is settled to some value, the tests waits to collect similar NRT data for a few times (default was 3).
Add the new environment variable E2E_NROP_COOLDOWN_THRESHOLD to enable the user to configure the number of times (that the collected resources data matches the initial) expected to consider the data is settled.

Add function to wait for the NRT data to settle based on the cooldown threshold, and replace the explicit call of the same wait with calls to the new function, this reduces coupling and increases cohesion between code parts.

Signed-off-by: shereenH <shajmakh@redhat.com>