e2e: remove unwanted skip
Seems like 3 tests were accedentally skipped in both scopes because of this check.
The removed check is pointeless because the nrtList is never empty because it
is collected in upper level BeforeEach. And even if it is, the tests should not
be skipped.

The intent of these tests is to do non-regression tests and check the numa-aware scheduler tolerate lack of NRT data for burstable and beset-effort QoS pods. Following this, let's also add a check to verify NRT status post scheduling the workloads.

Signed-off-by: shereenH <shajmakh@redhat.com>