e2e: serial: config: split infra and reference NRTList
The serial lane infra code needs a NRTList to make logic
and finalize its setup (run the numacell DS, most notably).
But we can't reuse this NRTList as reference point later on to
cooldown: it's recorded too early.

We should record the reference NRT list when all the infra setup
is done. So, we move the recording logic in its own method, called
at the right time (once all the infra setup succeeded) and
we split the duties of the infra NRT List (to be used only internally)
and the NRT list to be recorded and used later as reference.

Signed-off-by: Francesco Romani <fromani@redhat.com>