e2e: serial: infra: wait for numacell to be reported
Before to call the infra setup done, we need to wait
the numacell resources shows up *in NRT* (not just in the node objects).
If we fail to do that, we call infra succeeded too early,
and we have still a risk the reference NRT data is recorded
too early, without the numacell resources reported.

Signed-off-by: Francesco Romani <fromani@redhat.com>