e2e: serial: kconfig: support defaults
The kubelet sets defaults for Topology Manager policy and scope
when a value is not explicitely set in the configuration.
Let's support this scenario, copying the kubelet defaults.
We should get them programmatically and from the right
kubelet version, but the saving grace is the values don't change
frequently (if at all).

Signed-off-by: Francesco Romani <fromani@redhat.com>