serial: fetch CacheResyncPeriod from Status
When running the tests on a cluster with default configuration, the default value of scheduler's CacheResyncPeriod is reflected in the object's status and not in spec. The tests do not change this field so they run on default configuration and when fetching it from spec the value is not found hence skipping tests.

Fetch the scheduler's CacheResyncPeriod from its status instead to enable running the tests. This is also true if the value is not default.

Signed-off-by: shajmakh <shajmakh@redhat.com>