e2e: simplify runner
the e2e runner is one of the oldest parts still left

back in time we thought it could make sense to run the e2e against
a real cluster, but this option never materialized, and we ended up
using the runner only on CI.

When running against existing clusters, we run the specific suites
with extra parameters and config.

The end result is the runner script is unnecessary complex and actually
obfuscates the failure. Rewrite and simplify, tailoring for use on CI.

Signed-off-by: Francesco Romani <fromani@redhat.com>