rte:sysinfo, podrescompat removal:superseded by getallocatableResource
`sysinfo` and `podrescompat` packages were introduced as a way to obtain
resource information from the system and serve the endpoints provided by
podresources in the absence of getAllotableResource endpoint in podresource
API.

Support for getallocatableResource endpoint graduated to GA recently and
it is safe to assume that this endpoint is enabled by default and provides
all the necesary information for determining node resource allocatable.

We therefore no longer need the aforementioned packages and remove them.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>