e2e:rte: run local rte test on separate suite
The local rte tests are cluster-independent.

Although we're using ginkgo focus for running only the local tests,
the suite still tries to  initialize the code in the other tests.

The test passed so far only becuase of the fact that
k8s e2e framework knows how to deal with the absence of kubeconfig file.

Therefore, we should have a sperate suite for the local tests
which doesn't import other tests that are depending on the
presence of a cluster.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>