e2e: fix flaky test 47591
The test fails in CI and passes locally. The test modifies the pod resources under the deployment and do some checks on the new pod. The part that is causing the test to fail is that it doesn't wait "long enough" for the new pod to get created, causing the automation to pick an old pod, hence failing in later node name check.
 Investigating it deeper, even with same oc client the test kept passing locally. Yet managed eventually to reproduce it by running it using an ssh connection to the cluster. Not clear about the root cause at the moment, but it seems that when the automation is connected to the cluster actually matters in terms of how quick the check can be, as via exporting the KC the test had no chance of failing, but via ssh the test continuesly failed, possible indicator that ssh connection has a faster reaction than exporting KC.

To fix the flakiness, we add a waiting loop to guarantee we do not move further until the new pod's replicaset gets created.

Signed-off-by: shajmakh <shajmakh@redhat.com>