utils: fixture: improve debug print on failure
The cooldown is performed at the end of the test and its main goal is to validate that the NRT is reset to the initial value.
On reboot scenarios, this check always fails because it turned out that the memory distributaion accross the same numa zones of same node is not similar after the reboot, but is always found the same on node level. To illustrate that let's have a simple example:

NRT before reboot:

```
worker-0:
	numa-0:
	      - allocatable: "130674647040"
		available: "130674647040"
	        capacity: "135238049792"
	        name: memory
	      - allocatable: "39"
	        available: "39"
  	        capacity: "40"
	        name: cpu
        numa-1:
	      - allocatable: "130715820032"
                available: "130715820032"
                capacity: "135279222784"
                name: memory
              - allocatable: "39"
                available: "39"
                capacity: "40"
                name: cpu

```

NRT after reboot:

```
worker-0:
        numa-0:
              - allocatable: "130715820032"
                available: "130715820032"
                capacity: "135279222784"
                name: memory
              - allocatable: "39"
                available: "39"
                capacity: "40"
                name: cpu
        numa-1:
              - allocatable: "130674647040"
                available: "130674647040"
                capacity: "135238049792"
                name: memory
              - allocatable: "39"
                available: "39"
                capacity: "40"
                name: cpu
```

As one can see in the above example the node NRT overall is similar between before and after reboot, the only diffirence is that reboot may cause NRT to display switched alignment of the memory resources between both the numas.

The old way of presenting the difference between two NRTs was not helpful because the way we actually compare the NRTs is different and on which we decide if the NRTs are similar or not, as we sort the sources first then check them. However using cmp.Diff will display difference between the NRT items as they are without any sorting (which very likely to be always different) which is not a criterion to decide if the lists are similar.

Instead of displaying the difference, print out the initial and current NRT lists so we can debug the failure more clearly.

Fixing reboot cases follows in the next commit.

Signed-off-by: shajmakh <shajmakh@redhat.com>