serial: remove flaky test
This test revealed a race between the 2 schedulers, since we do not wait for the pod handled by TAS scheduler to get created, we cannot guarantee that it will pick up the resources before the pod that is handled by the default scheduler. because of this the test is fragile and fails frequently.
Finally, it was agreed to completely remove this test (option 1) for being too flaky and regardless how we change it, the change won't make it robust and will add little to no value.

Signed-off-by: shajmakh <shajmakh@redhat.com>