serial: skip tests before padding
4 tests from both TM scopes were skipped while the correct anticipation was to run. The reason they were skipped is because there were not enough NUMA nodes that could accommodate the required resources by the test's pod (needed 2 found 1). The way this info is checked in the tests is by checking that NRT has enough resources on the NUMAs and counts how many NUMAs can fit.

After investigating, it turns out that the tests pad the nodes before even verifying if the nodes' configuration in terms of TM policy and scope matches the test. Yet this is not a problem itself. The problem is that the tests later, after the padding is done, use utils. fixture.Skipf() which later on the teardown points out that there is no need to wait for NRTs to settle because the assumption is that the test called skipf did not trigger a change in the topologies, e.g create pods, while in this case it in fact did. So what happens is that the teardown clears the NS but without waiting enough time for the NRTs to settle, which leaves un-updated data for the next test in line.

To fix the root cause, we call the check of the TMscope and policy NUMA count before padding the nodes.

(note that we could also fix this by calling regular ginkgo.skip instead of fixture.skipf so at teardown it'll wait for the NRT settlement, but there is no point in padding the nodes for a test that should undoubtedly be skipped later)

Signed-off-by: shajmakh <shajmakh@redhat.com>