serial: replace requested device
The tests are intended for a device that is available only on one numa zone on a node. GetDeviceType1Name is deviceA which in our CI setup is available on all the numas of all nodes, thus causing the test to be skipped.

Replace the device name with GetDeviceType3Name which corresponds to deviceC that's configured only on one numa per compute node in order for these tests to run properly.

Signed-off-by: shajmakh <shajmakh@redhat.com>