serial: check node level NRT
The test creates 1 replicaset with 10 pods. and part of the pods are sometimes scheduled on different numas, and when the test was checking NRT data update, it was failing because the che
ck was done onuma level, whereas it should be done on node level.

Replace the NRT check to be done on node  situations where t replicas are scheduled on multi numa.

Signed-off-by: shajmakh <shajmakh@redhat.com>