controller: test: don't update objects  via status()
In test code, when updating a non-instance (e.g. NROP, NRSCHED)
object, we should use the main client update ('client.Update')
not the status subresource update which is meant, well, only
for status updates. This was likely sneaked in early on and gone
unnoticed because the fake client is too permissive,
but is going to blow on >= 1.27.

Signed-off-by: Francesco Romani <fromani@redhat.com>