e2e: serial: add test for related objs
Complement the current test we have for related objects with a e2e test
which actually queries the cluster to crosscheck the related objects.

To enable the test, move the related objects code to an internal
package. This code layour is however debatable and can be revisited
in the future.

Signed-off-by: Francesco Romani <fromani@redhat.com>