pkg: rte: fix unit tests
Previously if a flag was not supposed to be in the actual ds args the
unit tests were passing even if the flag existed in the actual ds. This is because
the test searched for the expected args within the actual ds args,
but not the opposite.

Make this check more accurate by comparing both expected and actual
flags which should be identical.

Signed-off-by: shajmakh <shajmakh@redhat.com>

TODO: solve rebase