sched: support for aggressive or relaxed resync
We configure the cache resync code to:
1. react to any foreign pod (= pod not scheduled by this scheduler)
2. consider all the pods when computing the expected node state

But thinking about it, only pods whose containers require exclusive
resources are relevant to the node state, because only exclusive
resources can have NUMA affinity.

We retroactively dub this configuration "aggressive" because it
causes recomputation when any change is detected, even unnecessary ones.

Recomputing the state often creates more work on the scheduler process,
but most notably slows down the scheduling because we need to wait more
for resyncs, being the current state invalidated (because of aggressive
detection).

To improve in both areas (resource consumption but most notably
scheduling latency) we add a new mode, called "relaxed" on which we
leverage the scheduler options to react and consider only pods whose
containers require exclusive resources.

The new "relaxed" mode is enabled by default, and the old behavior
can be re-enabled with a configuration option.

Signed-off-by: Francesco Romani <fromani@redhat.com>