e2e: serial: resacct: take snapshot after padding
Previously, the test fixture took its NRT snapshot before padding.
But padding is an essential component of testing, and requires
guaranteed class pods to actually consume numa-local resources.

If we later check that conditions match, we will never ever converge
because padding pods will be (correctly) counted to assess the node
resource consumption.

The correct approach is then to take the expected state AFTER padding,
considering padding pods as part of the cluster state, and right before
any test-specific pod is scheduled.

Or, alternatively, to check the state excluding the padding pods which
will make the test more complex (and thus fragile), and/or of
questionable value.

The only remaining question is why this failure was gone unnoticed
up until now.

Signed-off-by: Francesco Romani <fromani@redhat.com>