Set base CSV replaces field
If we release our newer operator version without the correct metadata, it will release that new version just fine but at the same time it will unpublish all earlier versions in the channel, making the new version effectively the only entry in that channel.

This happens when a new operator bundle ships with a "skipRange" annotation but without the "replaces" property in the operators OLM metadata (CSV)

This PR adds the "replaces" property in addition to the "skipRange" property.