chore: e2e: move filter into e2enrt util pkg
Move the filtering function in the utilities package.
This move is debatable, the main driver is this was the only
utility function that was missing from the utils pkg, so we
bring it in for the sake of consistency.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit 520cb7763a5655d51ab0df1254d2c91b78b5fde4)