hack: e2e: support hierarchical tiers in the runner
Enhance the runner script to support hierarchical tiers,
abusing --focus for now, labels in the future.

What is "hierarchical tiers"?
Currently running tier2 tests means we run ONLY tier2 tests, not tier1 and tier2 tests.
This is strictly speaking correct but is inconvenient.
Introducing hierarchy means that if we select a tier (and enable
hierarchy modes) then the tests which will run are the ones belonging
to that tier and to all higher tier)

IOW the selected tier will mean "run all tiers up to this one,
included".

Signed-off-by: Francesco Romani <fromani@redhat.com>