serial: tests: introduce tier 0
Now that we have a simple way to run all tests up to a given tier,
included (see: f69ac68a6375fec6ec355c44bd1c6d02a36a7593) we can
review our tiers.

The ter1 tests had 47 tests, roughly, which is great on a hand
but bad on the other hand, because the very purpose of tiering
was to have a strict and well delimited set of tests.

We promote the most critical tier1 tests to tier0 tests.
Tiering is not set in stone - tests will be moved among tiers.

But tier size should always kept at bay in the coming changes.
As rule of thumb, no tier should contain more than 20-25% of total
tests; and this is especially true for high (tier0/tier1) tests.

Signed-off-by: Francesco Romani <fromani@redhat.com>