api defaults: rename method
Defaults() can be missed with actually filling the empty resource object
with all default values. While the goal of this method is actually to
fill only the empty (nil) feilds in the resource instance.

Rename it to reflect to better reflect the intented functionality without the need to look into the
implementation.

Signed-off-by: shajmakh <shajmakh@redhat.com>