kubeletconfig: controller: add softowner labels
add labels to easily infer which MCP and which NROP
instance are linked to a generated configmap

The OwnerReference is already correctly set to
handle the cascading delete, we need more soft links
between objects and labels seems a good and natural fit.

We plan to use these labels for troubleshooting and
in the e2e tests.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit ada10950c5e2e1830048c45e4332b4656a7075f3)