sched: api: controller: enable sched HA
Add support in the scheduler controller to enable scheduler HA
using the tools provided by k8s framework already.
This basically boils down to set up the leader election and
the replica count properly.
We let the user control this exposing a new optional `Replicas`
value in the NUMAResourcesScheduler spec.

This commit adds code enablement and unit tests, e2e tests,
which are far from trivial, is deferred to future changes.
The existing e2e tests are expected to be good enough
to guarantee no regression.

Signed-off-by: Francesco Romani <fromani@redhat.com>