e2e: tolerations: test toleration on NROP CR creation
The new 2 tests (IDs: 72854, 72855) require no RTE pods to be
running. The only way ATM to achieve this is by not having nrop from the beginning. Considering this is part of serial, the cluster will have RTE running. So the only way to start clean again is by deleting NROP CR and creating it within the test scope per the test needs.

The setup is similar to both which is requiring to taint one worker node before creating the new NROP CR.
The two tests complement one another as the first verifies the
functionality after adding tolerations to existing NROP CR. While the
second test verifies that the tolerations, added to the NROP CR prior
to creating it, would be processed correctly.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>