e2e: wait for ds pods creation at teardown
The problem with the current ds teardown is that it waits for the ds to
be ready without waiting for pods count in parallel. This is problematic
because at some scenarios the ds even with fewer pods than expected at teardown, it will still report "ready" but will fail later when actually
checking the pods' count.

To fix this, we add another wait loop that waits for the creation of the
expected pods count for the ds before verifying it is ready.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>