internal: return nil as error to keep the loop going
PollUntilContextTimeout(..) will return if the returned error is not nil
or if the condition is met. Specefically in ForDaemonsetPodsCreation(),
there is a chance that the Ds is still not created so instead of waiting
for the Ds to be created seperately allow the loop to keep waiting even
when the api fails to get the Ds from the cluster.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>