images: streamline overhaul the image management
We bundle operator and operand (RTE) in the same image.
The operator learn which is its own image by inspecting
its own pod, and the user can override the value setting
the corresponding field in the NodeConfig struct.

Since the beginning we wanted also to make the images
configurable for development purposes. This feature
never materialized.

In addition, the detection and image replacement code
got messier over time, with no real cleanup.

To address both these needs, overhaul and simplify the
image management. All changes are expected to
be transparent for regular users, outside development
environments.

Signed-off-by: Francesco Romani <fromani@redhat.com>