e2e serial: initiate NRT list in each test independently
Before this change, each test used the data of NRT list that was
gathered at the beginning of the suite and compared it with the NRT data gathered after each test to ensure the NRTs are settled. This is fine as long as the resources topology is guaranteed not to be updated in a way that messes up the teardown comparison.

Since we cannot specify the tools used to set up the resources in the
testing environment, hence no guarantee the topology would stay the same from the beginning of the suite till the end; we record the NRT data on each test fetching it directly from the cluster, independent from the data gathered at the start of the suite.

Also add a warning at the test setup if NRT data was mismatching the
data gathered at the beginning of the suite.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>