utils: add a function to compare NRT lists items
Add a function to verify if two NRT lists have equal items in terms of
nodes, zones and resources amounts.

Note: this was added under utils and not under internal because it
depends on helper utils functions so moving it to internal would create
import cycle; and because the dependencies functions are not tested.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>