e2e: drop host-level resources when padding numas
A problem was caught during workload placement tests requiring pre-padding numa zones as test setup. The problem was that the cluster
had leftover workloads that consumed host-level resources and that was
counted in the base load of the node. Later on, the test expected to pad
Numa zones consider host-level resources to be part of the padder
resources. The setup failed because host-level resources are not numa
scoped.

Adjust the relevant methods that pad numa zones to filter specified
resources such as ephemeral storage.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>