controller: reconcile: split functions
split the nrop/controller `reconcileResources` in steps.
Each step manages the spec/status pair of the machineconfig
or of the daemonset, which are the major resources managed
by this controller.

We intentionally split the function respecting the current layout.
A better refactoring will probably reshape most of the controller
and of the helper functions, but this seems sufficient for now.

The benefit of this cleanup is to enable future changes
like more modular reconciliation and, in the more distant future,
easier removal of machineconfig/selinux step.

Signed-off-by: Francesco Romani <fromani@redhat.com>