e2e: hostlevel: wait more while pod in pending state
Apperantly the fix in #944 was not enough to solve the issue completely.
ALthough it helped in showing the events for debugging, it wasn't enough for
checking the events as sometimes the events count would still be 0 which
then fails the test immediately.

The pending phase of a pod indicates that it has been accepted by the Kubernetes system, but it is not yet scheduled to run on a node.
Thus it is very possible that during the events check, the pod was still
being processed and not yet handled by the scheduler hence no events
yet.

To fix this, we wait more time while pod is in Pending state to make
sure that all pod events are logged.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>