e2e: rearrange defer funcs
The old version had a defer block that did all the cleanup at once, but
defer statements should be placed near to the point they are related for
better readability and troubleshooting.

Split the old defer statement into smaller ones to avoid heavy
proccessing and move them to the nearest point where they relate.

Note: multiple defer statements in a function are executed in
last-in-first-out (LIFO) order.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>