e2e: label tests for non supporting versions
For running the regressions we use the latest tests image (from master), with no filters on the tests so all tests executed on all versions.

The problem is that some tests are relevant only for specific releases and executing them on non-supporting versions results in greater execution time and noisy failures.

Considering so far that there are no major changes in the operator
functionality between the different supported releases, the proposed
update is to label specs that aren't relevant to run on specific ystream,
and in the d/s CI (as serial doesn't run u/s) we use
--ginkgo.label-filter=QUERY to filter out these tests.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>