[release-4.15] controller: switch back defaults
It's better to not change defaults in a stable branch unless
the benefits are overwhelming and the risks minimal. There's
strong hints this is the case, but no strong evidence yet.

So we revert the defaults to their original values.
Effectively, this series enables the support but doesn't
enable by default.

To have all the improvements enabled by default, this change
needs to be reverted and
e6582903af1d5552d9d56961d481eee5d6d5aae6
needs to be applied.

Signed-off-by: Francesco Romani <fromani@redhat.com>