makefile: build serial bin with race detector
Tests in serial often fails and there is no clear reason from the logs,
although elaborated, on why the failure happened.

Build the serial suite binary with race detector so that when tests run
a warning is shown pointing to occured race.

This is meant to help debug the failing tests but not actually fixing them or preventing race
conditions.

xref: https://go.dev/doc/articles/race_detector

Signed-off-by: Shereen Haj <shajmakh@redhat.com>