e2e: rmsched -> schedrst
Tests that restart the scheduler actually remove the scheduler pod but
not the scheduler, and tests that remove the scheduler CR and recover are restarting the
scheduler.

Both versions delete the scheduler logs which are essential for
troubleshooting if needed.

Replace the `rmsched` tag with a tag that better reflects the effect of tests under this go file.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>