e2e: fail on errgroup.Wait() error
The function was reporting the error without failing and the test would
proceed to execute instead of recovering and exit.

Update the function to return the result of eg.Wait() and handle the
error outside.

Note this does not fix the error cause so that should be handled in
another commit.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>