e2e: adapt performance profile existance
 Considering that d/s ci runs on env with PP configured
according to the RDS, when the test updates kubeletconfig directly, which is owned by another object, the update is not applied; hence, the test fails.

Extend the test to detect if the kubeletconfig belongs to a
performanceprofile, and if so update the performance profile instead so that the updates actually apply.

This is the first time NTO dependencies are needed so also update the
dependencies. Note that the NTO version used in the dependencies is from 4.16 branch because starting 4.17 NTO is consuming Go 1.22, and using the latest NTO will require bumping the Go version for NROP as well, which is something we are not clear on yet, thus sticking with 1.21, since all we need is just use performance profile type.

So far the serial suite has 2 tests that apply update to the
kubeletconfig. This PR targets one test (id-47585), another PR for
the other test will follow.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>