e2e: enhance MCP update check
In extreme cases where the mcp update happens too fast, some test checks can be a false positive but because the check happens on pulses it may miss the status of the condition that the test is looking for.

The approach to solving this flakiness is first summarized in below
steps:

`
1. save the initial config under the mcp that would be updated
2. issue the event to trigger mcp update
3. loop until the below conditions are met (&&):
 3.1  condition Updated=true
 3.2  a random node under the updating mcp has config
desired==current==newExpectedOnMcp
`

The problem is that it may also be flaky when mcp conditions
and node annotations are not updated immediately after the causing event and in that case, the loop would pass from the first iteration.

So we continue with the approach, but we add a meaningful wait before verifying 3.1. and that is to wait until Updating==true. We make that a soft requirement though because in some cases the switch
Updated->Updating->Updated can also be faster than the loop intervals
(even though they are reduced). This will guarantee that the test waited
enough time to allow the update to start.

so we are at this point:

`
1. save the initial config under the mcp that would be updated
2. issue the event to trigger mcp update
3. soft requirement: loop over until condition Updating==true
4. loop until the below conditions are met (&&):
 4.1  condition Updated=true
 4.2  a random node under the updating mcp has config
desired==current==newExpectedOnMcp
`

We adjust the algorithm a bit more and remove 4.2 from the loop. The
reason is that 4.2 is a result of 4.1 in a loop. It would be reasonable
if it fails when Updated==true, and a failure would make sense only if
the loop for Updated is completed.

So we end up with the below algorithm version:
1. save the initial config under the mcp that would be updated
2. issue the event to trigger mcp update
3. soft requirement: loop over until condition Updating==true
4. loop until condition Updated=true
5. a sample node under the updating mcp has config
desired==current==newExpectedOnMcp
`

Signed-off-by: Shereen Haj <shajmakh@redhat.com>