relaxation for the error because only one pod should be pending
Here I provided a relaxation for the error. The reason is because the test is sometimes flaky because the time frame needed
after the kubeletconfig changes to creating the deployment needed a few more seconds after the update to pick up the new changes and recover,
hence can cause the pods list length to be more than one with one pod that is pending and several other pods that are in ContainerStatusUnknown
which have zero effect but will fail the test.
So as long as there is exactly one pod that is pending the scheduler has behaved correctly.

Signed-off-by: Roy Shemtov <rshemtov@redhat.com>