e2e: adapt performance profile existence and stabalize test
This PR follows #962 for adding performance profile dependencies and it adds detection for kubeletconfig changes through performance profile. Also this PR adds test id to the test 75354.
This is a reboot test that is changing the kubeletconfig configuration. The changes here intend to detect if a performance profile exists
and if it does to change the kubeletconfig changes through it. If the performance profile exists and these changes were not in place the test would fail.
So with these changes the test will be smarter in picking where kubeletconfig changes needs to be, from the kubeletconfig itself or the performance profile.

Signed-off-by: Roy Shemtov <rshemtov@redhat.com>