e2e: relaxing and stabalizing the test case
The reason for this commit are for relaxing and stabalizing the test case as follows.
The test case at the change is asserting for the length of the pods to be equal to 1. This assertion results in a flaky behavior because before that assertion we have made changes to the kubeletconfig and waited for the reboot to take place, after the reboot we immediately create a deployment with the expectation to have exactly one pending pod. The issue arrises when sometimes there are more than 1 pod which is not expected and than the test fails.
It appears that the correct behavior is right for just a single pod which is correctly pending but there are other pods that their status is ContainerStatusUnknown,
This happens because the deployment was created immediately after the kubeletchanges and it needed an additonal seconds before it picked up the new changes and recovered, This makes the other pods with ContainerStatusUnknown have zero affect on the test case because as long as there is exactly one pod that is pending the test should pass.

Signed-off-by: Roy Shemtov <rshemtov@redhat.com>