verify-generated: do no compare `createdAt` field
The `createdAt` field in the CSV file always gets updated
when running `make bundle` even
if nothing else has changed in the CSV file.

In order to avoid these false positives,
the verification script will dump the CSV file if
the only change is at the `createdAt` line in the CSV file.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>