controller: introduce kubeletConfigHandler object
The business logic of kubelet controller is similar when running on OCP or HCP.
However there are slighty different parameters that some of the
existing functions in the reconcile loop accepts.

We introduce the `kubeletConfigHandler` so we could maintain a common
API for the functions for both OCP and HCP platforms.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>